change architecture after mentoring
