+++ src/main/java/com/tdd/hhplus_tdd_2week_java/domain/applied_lecture/AppliedLectureRepository.java
@@ -1,0 +1,6 @@
+package com.tdd.hhplus_tdd_2week_java.domain.applied_lecture;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AppliedLectureRepository extends JpaRepository<AppliedLecture, Long> {
+}
