+++ src/main/java/com/tdd/hhplus_tdd_2week_java/domain/lecture/LectureRepository.java
@@ -1,7 +1,8 @@
-package com.tdd.hhplus_tdd_2week_java.repository.lecture;
+package com.tdd.hhplus_tdd_2week_java.domain.lecture;
 
-import com.tdd.hhplus_tdd_2week_java.entity.Lecture;
+import com.tdd.hhplus_tdd_2week_java.infrastructure.lecture.LectureRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LectureRepository extends JpaRepository<Lecture, Long>, LectureRepositoryCustom {
+
 }
