+++ src/main/java/com/tdd/hhplus_tdd_2week_java/domain/studuent/StudentRepository.java
@@ -1,0 +1,6 @@
+package com.tdd.hhplus_tdd_2week_java.domain.studuent;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudentRepository extends JpaRepository<Student, Long> {
+}
