--- src/main/java/com/tdd/hhplus_tdd_2week_java/entity/AppliedLecture.java
@@ -1,26 +1,0 @@
-package com.tdd.hhplus_tdd_2week_java.entity;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
-@Table(name = "applied_lecture")
-public class AppliedLecture extends BaseEntity{
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "lecture")
-    private Lecture lecture;
-
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "student")
-    private Student student;
-
-
-
-}
