+++ src/main/java/com/tdd/hhplus_tdd_2week_java/infrastructure/applied_lecture/AppliedLectureRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package com.tdd.hhplus_tdd_2week_java.repository.applied_lecture;
+package com.tdd.hhplus_tdd_2week_java.infrastructure.applied_lecture;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
@@ -6,6 +6,6 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @RequiredArgsConstructor
-public class AppliedLectureRepositoryCustomImpl implements  AppliedLectureRepositoryCustom {
+public class AppliedLectureRepositoryCustomImpl {
     private final JPAQueryFactory dsl;
 }
