+++ src/main/java/com/tdd/hhplus_tdd_2week_java/infrastructure/lecture/LectureRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.tdd.hhplus_tdd_2week_java.infrastructure.lecture;
+
+public interface LectureRepositoryCustom {
+    void findByLecturePid();
+}
