+++ src/main/java/com/tdd/hhplus_tdd_2week_java/infrastructure/student/StudentRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.tdd.hhplus_tdd_2week_java.infrastructure.student;
+
+public interface StudentRepositoryCustom {
+}
