+++ src/main/java/com/tdd/hhplus_tdd_2week_java/infrastructure/student/StudentRepositoryCustomImpl.java
@@ -1,0 +1,4 @@
+package com.tdd.hhplus_tdd_2week_java.infrastructure.student;
+
+public class StudentRepositoryCustomImpl implements StudentRepositoryCustom{
+}
