--- src/main/java/com/tdd/hhplus_tdd_2week_java/repository/lecture/LectureRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.tdd.hhplus_tdd_2week_java.repository.lecture;
-
-public interface LectureRepositoryCustom {
-}
