--- src/main/java/com/tdd/hhplus_tdd_2week_java/repository/student/StudentRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.tdd.hhplus_tdd_2week_java.repository.student;
-
-import com.tdd.hhplus_tdd_2week_java.entity.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentRepositoryCustom  {
-}
