--- src/main/java/com/tdd/hhplus_tdd_2week_java/service/student/StudentService.java
@@ -1,7 +1,0 @@
-package com.tdd.hhplus_tdd_2week_java.service.student;
-
-
-import com.tdd.hhplus_tdd_2week_java.entity.Student;
-
-public interface StudentService {
-}
